apps/f1-car-data-stub/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "f1_car_data_stub" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/apps/f1-car-data-stub/html'
 --cov-report xml:'../../coverage/apps/f1-car-data-stub/coverage.xml'
 --html='../../reports/apps/f1-car-data-stub/unittests/html/index.html'
 --junitxml='../../reports/apps/f1-car-data-stub/unittests/junit.xml'
"""

[tool.poetry]
name = "f1-car-data-stub"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "f1_car_data_stub"

  [tool.poetry.dependencies]
  python = ">=3.9"
  pymap3d = "3.1.0"
  fastf1 = "3.4.3"
  kafka-python = "2.2.4"
  colorlog = "6.9.0"

    [tool.poetry.dependencies.f1-replay-data-model-py]
    path = "../../libs/f1-replay-data-model/python"
    develop = true

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
